game: Move game part to an internal package
